+++ test/e2e/modules/views/outputs.tf
@@ -1,0 +1,4 @@
+output "view" {
+  description = "The view the network resolves to."
+  value       = powerdns_network.internal.view
+}